- WinRT window scaling fixes

+++ src/windowsystems/WinRT/WinRT_Window.h
@@ -55,7 +55,7 @@ namespace april
 		void terminateMainLoop();
 		void showVirtualKeyboard();
 		void hideVirtualKeyboard();
-		void changeSize(int w, int h); // required override instead of normal size changing
+		void changeSize(float w, float h); // required override instead of normal size changing
 		
 	protected:
 		int width;
